# src/main/java/com/boxboat/simpletracker/api/ContractsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.boxboat.simpletracker.api;

import com.boxboat.simpletracker.model.Contract;
import com.boxboat.simpletracker.model.ContractPatch;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-03-07T14:16:30.773327-08:00[America/Los_Angeles]")
@Validated
@Tag(name = "contracts", description = "the contracts API")
public interface ContractsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * DELETE /contracts/{contractId} : Delete a contract by ID
     *
     * @param contractId Numeric ID of the contract (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "contractsContractIdDelete",
        summary = "Delete a contract by ID",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/contracts/{contractId}"
    )
    default ResponseEntity<Void> contractsContractIdDelete(
        @Parameter(name = "contractId", description = "Numeric ID of the contract", required = true, in = ParameterIn.PATH) @PathVariable("contractId") Integer contractId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /contracts/{contractId} : Get a contract by ID
     *
     * @param contractId Numeric ID of the contract (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "contractsContractIdGet",
        summary = "Get a contract by ID",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Contract.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/contracts/{contractId}",
        produces = { "application/json" }
    )
    default ResponseEntity<Contract> contractsContractIdGet(
        @Parameter(name = "contractId", description = "Numeric ID of the contract", required = true, in = ParameterIn.PATH) @PathVariable("contractId") Integer contractId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"tech\" : [ \"tech\", \"tech\" ], \"clientId\" : 6, \"endDate\" : \"endDate\", \"id\" : 0, \"type\" : \"type\", \"startDate\" : \"startDate\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /contracts/{contractId} : Update a contract by ID
     *
     * @param contractId Numeric ID of the contract (required)
     * @param contractPatch  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "contractsContractIdPut",
        summary = "Update a contract by ID",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Contract.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/contracts/{contractId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Contract> contractsContractIdPut(
        @Parameter(name = "contractId", description = "Numeric ID of the contract", required = true, in = ParameterIn.PATH) @PathVariable("contractId") Integer contractId,
        @Parameter(name = "ContractPatch", description = "", required = true) @Valid @RequestBody ContractPatch contractPatch
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"tech\" : [ \"tech\", \"tech\" ], \"clientId\" : 6, \"endDate\" : \"endDate\", \"id\" : 0, \"type\" : \"type\", \"startDate\" : \"startDate\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /contracts : Get all contracts
     * Retrieves all contracts
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "contractsGet",
        summary = "Get all contracts",
        description = "Retrieves all contracts",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Contract.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/contracts",
        produces = { "application/json" }
    )
    default ResponseEntity<List<Contract>> contractsGet(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"tech\" : [ \"tech\", \"tech\" ], \"clientId\" : 6, \"endDate\" : \"endDate\", \"id\" : 0, \"type\" : \"type\", \"startDate\" : \"startDate\" }, { \"tech\" : [ \"tech\", \"tech\" ], \"clientId\" : 6, \"endDate\" : \"endDate\", \"id\" : 0, \"type\" : \"type\", \"startDate\" : \"startDate\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /contracts : Create new contract
     * Add a new contract with fields
     *
     * @param contractPatch  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "contractsPost",
        summary = "Create new contract",
        description = "Add a new contract with fields",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Contract.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/contracts",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Contract> contractsPost(
        @Parameter(name = "ContractPatch", description = "", required = true) @Valid @RequestBody ContractPatch contractPatch
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"tech\" : [ \"tech\", \"tech\" ], \"clientId\" : 6, \"endDate\" : \"endDate\", \"id\" : 0, \"type\" : \"type\", \"startDate\" : \"startDate\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}