# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  description: |
    This API allows management of client contracts and tracks history of employee contract assignments
  title: Simple emlpoyee contract tracker api
  version: 1.0.0
servers:
- url: /
paths:
...
  /clients:
    get:
      description: Retrieves all clients
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/client'
                type: array
          description: OK
      summary: Get all clients
      x-accepts: application/json
    post:
      description: Add a new client with given name and url
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_clients_post_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/client'
          description: OK
      summary: Create new client
      x-content-type: application/json
      x-accepts: application/json
  /clients/{clientId}:
    delete:
      parameters:
      - description: Numeric ID of the client
        explode: false
        in: path
        name: clientId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      summary: Delete a client by ID
      x-accepts: application/json
    get:
      parameters:
      - description: Numeric ID of the client
        explode: false
        in: path
        name: clientId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/client'
          description: OK
      summary: Get a client by ID
      x-accepts: application/json
    put:
      parameters:
      - description: Numeric ID of the client
        explode: false
        in: path
        name: clientId
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_clients_post_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/client'
          description: OK
      summary: Update a client by ID
      x-content-type: application/json
      x-accepts: application/json
  /contracts:
    get:
      description: Retrieves all contracts
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/contract'
                type: array
          description: OK
      summary: Get all contracts
      x-accepts: application/json
    post:
      description: Add a new contract with fields
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/contractPatch'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contract'
          description: OK
      summary: Create new contract
      x-content-type: application/json
      x-accepts: application/json
  /contracts/{contractId}:
    delete:
      parameters:
      - description: Numeric ID of the contract
        explode: false
        in: path
        name: contractId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      summary: Delete a contract by ID
      x-accepts: application/json
    get:
      parameters:
      - description: Numeric ID of the contract
        explode: false
        in: path
        name: contractId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contract'
          description: OK
      summary: Get a contract by ID
      x-accepts: application/json
    put:
      parameters:
      - description: Numeric ID of the contract
        explode: false
        in: path
        name: contractId
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/contractPatch'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contract'
          description: OK
      summary: Update a contract by ID
      x-content-type: application/json
      x-accepts: application/json
...
components:
...
  schemas:
...
    client:
      description: Client data
      example:
        name: name
        id: 0
        url: url
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
      type: object
    contract:
      description: Contract data
      example:
        tech:
        - tech
        - tech
        clientId: 6
        endDate: endDate
        id: 0
        type: type
        startDate: startDate
      properties:
        id:
          type: integer
        clientId:
          type: integer
        type:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        tech:
          items:
            type: string
          type: array
      type: object
    contractPatch:
      description: Contract data for update or creation
      properties:
        clientId:
          type: integer
        type:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        tech:
          items:
            type: string
          type: array
      type: object
...
    _clients_post_request:
      properties:
        name:
          type: string
        url:
          type: string
      type: object